PLEASE NOTE: This revision may very well break existing code due to a core change, read below for details and instructions. This code has not yet been tested, it is committed for the purpose of allowing Bantu and the team to work with this new code.
- The titania class now assigns the phpBB variables: $auth, $config, $db, $template, $user and $cache to the static variables which can be accessed using: titania::$db, titania::$template for example.
The titania class assigns these values in the static construct() method, but this would likely need to be moved to the initialise method OR the static construct method would need to be called in the common.php file to initialise the globals.
The reason for this change means that we will no longer have to globalise any phpBB variables any more.

There is likely a conflict with the $config member within the static class due to the phpBB $config variable being used now. Consider alternatives for the Titania specific config or better yet... merge the titania config data into the titania::$config array.

- Changed the trigger_error method a little to update it to allow a custom title and default to E_USER_NOTICE which is the default.
- Vastly improved error_box, we now assign the errors to 'errorbox' template block array, and success messages to the 'successbox' template block array. This way success messages should appear under the 'green' box each separated by a new line, and errors appear under a red box each separated by a new line.