Sorry, this is breaking a lot again :( See: http://www.phpbb.com/community/viewtopic.php?f=89&t=1647765
Need to uninstall and reinstall the database again (don't forget to remove the extra modules/tables left over).  Hopefully this will be the last time we need to uninstall/reinstall, the rest should work in updates.

Removed a bunch of language stuff from common (and others).  Lets re-add as needed and put in the appropriate file this time (ex: contrib details stuff does not belong in common).

Category list now works/displays on the index contrib list *should* work too, but not tested yet.