Changed some things around to make the objects more standardized.
generate_text_for_edit now returns the array from generate_text_for_edit (subject and options merged in as well) so our message tool can be used much more easily.

Lots of stuff changed in the FAQ.  Set up the new/edit entry to use the same message tool I built for the posts backend.  Also cleaned up the styles a bit.  Still a few things that need to be done for that yet though, like the folder image for the list display and the smilies are not working for some reason.  Added a validate function for the FAQ and cleaned up some unnecessary functions.php