Well, this is a pretty major update to the attachments system.
Re-worked how attachments are uploaded in the upload.php file.

If you guys can manage to get this working on your local host, please mess around with it and let me know what you think. I know there is a lot of stuff that could be changed and this is still under development so it is in no way close to being complete.

Note that when you create a contrib you will not have to upload an download until you want, I have just placed it there for testing. Currently you also have to have javascript enabled to upload files and only zip archives are allowed.