Few minor changes.
Remove the load_object and similar function calls in a few more places. Still need to clean them all up in some files.

The author object method, assign_details, now accepts a row to assign to the template again for the coauthor or extra authors, however it can now revert back to the old object data so nothing is lost and no unexpected results occur. This saves us from creating a duplicate object for the coauthor.