Another bug - not setting the last post info on topic creation
Correct reply count for topic list

Have to change the post_data functions and message class a bit as we need to check the auth settings once in a while (updating all post_data functions to use the same standard).  Fixing some bugs and missing auth checks related to this.