Huge commit here...
Lots of work on attachments and revisions.
Moving revisions to the Manage Contribution Page.
No more uploadify, we are going a simpler route that doesn't give us the problems uploadify did (especially security related)
Still some more work to do, but I don't want to sit on these changes forever (and have to kill someone if the hard drive crashed)

PLEASE REINSTALL THE DATABASE SECTIONS WITH THE INSTALLER.  The extensions have been changed majorly and attachments won't work without it.