NOTE: This update requires reinstallation of the database if you have any posts anywhere.
Rather significant change in how posts and topics are handled.

Decoupling them from contributions more as it posed issues with the queue.  Now storing topic_url and post_url to hold the unbuilt url location of the post or topic parent.  Renaming the contrib_id field to parent_id for topics.

This also saves us some queries as we don't have to generate the url on the fly (at least that we don't need the contribution name and type to generate them).