Truncate attention table on conversion
Redid some of how the tracking was setup so we are now able to much more easily track multiple things (such as marking all items in one area as read).

I've tested the mark topics read quite a bit, but it's pretty complex and it's possible that it is buggy.