Nearly finished the Category Management system at this stage of the game. Creating new categories and editing existing categories works well, except that newly-created categories are not assigned left_id and right_id values in the database yet, which will cause problems! - I will get to that tomorrow.
Delete category function should not be used as of yet, as it is still buggy and has not been worked on since the last commit; it will be worked on last. Like I said, some things are buggy with the system because of the lack of left_id and right_id values for new categories. Everything else checks out OK.