Extracting the user selection from the queries when grabbing the contribs.
Errors occur when a user did not exist before this (could happen if the user is deleted, titania isn't notified by phpBB when that is done and we'd have to edit code otherwise).  Now if the user doesn't exist they are shown as anonymous.

Some optimizations (removing unneeded SQL queries)