Do not create objects in loops unless we have to.
Just noticed this join query selects the category id based on the contrib type, which will not work for anything besides the vanilla install.

Changing the link to the contribution name to go to the contribution and adding a discussion/support link

Do NOT commit without viewing the diff first.  Reverting changes someone else has made on accident is not acceptable and happens far too often.