Fixing a few bugs if use_queue is disabled.
build_clean_url function in the url class (removes the sid after building) and (should) fix #62112

Moving the update release topic code into the contrib class.  Keeping 90% duplicate code in three areas is making things difficult to fix and change it.

Update the release topic when the contribution is updated #62116