[ticket/62550] Correctly build category-url
Currently the "categoryname-id"-component is added to the end of the url. But
this conflicts with the behavior of the sid, which leads to a new sessions. So
you get logged out with cookies disabled. The fix is quite easy. The
"categoryname-id"-component needs to be part of the url or url_building
function and not be appended afterwards.